% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{add_attr}
\alias{add_attr}
\title{Add attribute of quadratic model}
\usage{
add_attr(x)
}
\arguments{
\item{x}{a \code{\link{tbl_brain}}}
}
\description{
This function add attribute to tbl_brain data on xy plane, yz plane and xz plane. This facilitate users to
access the model information of the data.
}
\examples{
ro_brain_att <- add_attr(ro_brain)

}
