% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_pkg}
\alias{add_pkg}
\title{Add Packages to a Repository}
\usage{
add_pkg(pkg, repo = get_repo_location(), type = "source",
  overwrite = FALSE)
}
\arguments{
\item{pkg}{A vector of package bundles to add, e.g. \code{"cranium.tar.gz"}.}

\item{repo}{The location of the package repository.}

\item{type}{The type of the package. Only \code{"source"} is supported.}

\item{overwrite}{When \code{TRUE}, overwrite any existing packages that match
the input bundles.}
}
\value{
A logical vector indicating which packages were added successfully.
}
\description{
Add Packages to a Repository
}
