\name{add_quotes}
\alias{add_quotes}
\title{Add Quotes}
\usage{
  add_quotes(str)
}
\arguments{
  \item{str}{character string}
}
\value{
  quoted character string
}
\description{
  Adds double quotation marks around a character string
}
\details{
  This function is used to create the JSON file for package
  dependencies and reverse dependencies
}
\examples{
\dontrun{
  s = 'quotations'
  add_quotes(s)
}
}
\keyword{internal}

