% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_pkg}
\alias{archive_pkg}
\title{Move Packages to the Repository's Archive}
\usage{
archive_pkg(pkg, version, repo = get_repo_location(), type = "source",
  keep = TRUE, overwrite = FALSE)
}
\arguments{
\item{pkg}{A vector of package names to archive, e.g. \code{"cranium"}.}

\item{version}{A vector of valid package versions to archive.}

\item{repo}{The location of the package repository.}

\item{type}{The type of the package. Only \code{"source"} is supported.}

\item{keep}{When \code{TRUE}, do not remove the package from the repository
after it has been archived.}

\item{overwrite}{When \code{TRUE}, overwrite existing archived packages, if
present.}
}
\value{
A logical vector indicating which packages were moved to the archive
  successfully.
}
\description{
Move Packages to the Repository's Archive
}
