% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{change_coordinates}
\alias{change_coordinates}
\title{Change to parabolic coordinates}
\usage{
change_coordinates(obj, ...)
}
\arguments{
\item{obj}{an object}

\item{...}{currently ignored}
}
\description{
Change to parabolic coordinates
}
\examples{

file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  plot3d(tidy_brain)
  tidy_flat <- reorient(tidy_brain)
  plot3d(tidy_flat)
  spherical <- change_coordinates(tidy_flat)
  summary(spherical)
  }
}
}
