% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{clean}
\alias{clean}
\title{Filter a brain image by threshold}
\usage{
clean(x, threshold = 0.99, ...)
}
\arguments{
\item{x}{A \code{\link{tbl_brain}} object}

\item{threshold}{value below which points will not be plotted}

\item{...}{currently ignored}
}
\description{
Filter a brain image by threshold
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  nrow(tidy_brain)
  clean(tidy_brain) \%>\%
    nrow()
}
}
}
