% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{create_ro_matrix}
\alias{create_ro_matrix}
\title{create rotation matrix}
\usage{
create_ro_matrix(rotation_angle_radian, axis = "x")
}
\arguments{
\item{rotation_angle_radian}{radian of the estimated rotation angle}

\item{axis}{the axis that you are rotating around}
}
\description{
creat 3 by 3 rotation matrix for error B correction
}
\examples{
degree = 45
radian = deg2rad(degree)
ro_matrix <- create_ro_matrix(radian, axis = "y")
}
