\name{current_pkg_url}
\alias{current_pkg_url}
\title{URL of a package current description}
\usage{
  current_pkg_url(package)
}
\arguments{
  \item{package}{name of an R package in CRAN}
}
\value{
  complete URL of current package description
}
\description{
  Returns the URL of a package current description in CRAN
}
\examples{
\dontrun{
  # URL of current description of 'plspm'
  current_pkg_url("plspm")
}
}
\keyword{internal}

