% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_brains}
\alias{download_brains}
\alias{read_brains}
\title{Download YouToo sample data}
\usage{
download_brains(path = ".", type = "wildtype", pattern = NULL, ...)

read_brains(file)
}
\arguments{
\item{path}{where downloaded data will be stored}

\item{type}{\code{wildtype} or \code{youtoo}. Default is \code{wildtype}}

\item{pattern}{regular expression to match file names}

\item{...}{arguments passed to \code{\link{download.file}}}

\item{file}{path to HDF5 file}
}
\value{
a list of class \code{brain}
}
\description{
Download YouToo sample data
}
\examples{
wildtype_files <- download_brains(pattern = "101")
youtoo_files <- download_brains(type = "youtoo", pattern = "6")
brains <- read_brains(c(wildtype_files, youtoo_files))
plot2d(brains[[1]], title = name)
}
