% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_youtoo_data}
\alias{download_youtoo_data}
\alias{download_wildtype_data}
\title{Download You Too sample data}
\usage{
download_youtoo_data(folder)

download_wildtype_data(folder)
}
\arguments{
\item{folder}{where downloaded data will be stored}
}
\value{
a list of class \code{brain}
}
\description{
Download You Too sample data
}
\examples{
youtoo <- download_youtoo_data('folder')
plot2d(youtoo[[1]], title=name)
}
