% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{extract_description}
\alias{extract_description}
\title{Extract description from source package(s)}
\usage{
extract_description(x)
}
\arguments{
\item{x}{a .tar.gz package}
}
\value{
a data.table/data.frame object containing Description info
}
\description{
Extract description from source package(s)
}
\examples{
extract_description('test.tar.gz')
}

