% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{get_jawbone}
\alias{get_jawbone}
\title{Compute the jawbone model}
\usage{
get_jawbone(xyz, ...)
}
\arguments{
\item{xyz}{a \code{\link{tbl_brain}}}

\item{...}{currently ignored}
}
\description{
Compute the jawbone model
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
brain <- read_h5(file)
jbone <- get_jawbone(tidy(brain))
}
}
