% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_packages_fields}
\alias{get_packages_fields}
\title{Get Current package fields}
\usage{
get_packages_fields()
}
\value{
current fields used in the PACKAGES file
}
\description{
Get Current package fields
}
