\name{get_pkg_attribute}
\alias{get_pkg_attribute}
\alias{get_pkg_author}
\alias{get_pkg_citation}
\alias{get_pkg_compilation}
\alias{get_pkg_date}
\alias{get_pkg_depends}
\alias{get_pkg_imports}
\alias{get_pkg_license}
\alias{get_pkg_maintainer}
\alias{get_pkg_materials}
\alias{get_pkg_rev_depends}
\alias{get_pkg_rev_imports}
\alias{get_pkg_rev_suggests}
\alias{get_pkg_suggests}
\alias{get_pkg_url}
\alias{get_pkg_version}
\alias{get_pkg_views}
\title{Get Package Attribute}
\usage{
  get_pkg_attribute(pkg_attrs, attribute)
}
\arguments{
  \item{pkg_attrs}{data frame with package attributes}

  \item{attribute}{the name of an attribute}
}
\description{
  Extracts a given package attribute
}
\keyword{internal}

