\name{get_pkg_description}
\alias{get_pkg_description}
\title{Get Package Description}
\usage{
  get_pkg_description(pkg_doc)
}
\arguments{
  \item{pkg_doc}{an object of class
  \code{"HTMLInternalDocument"}}
}
\description{
  Extracts the package description from parsed html content
}
\keyword{internal}

