\name{get_table_reverse}
\alias{get_table_reverse}
\title{Get Table of Reverse Dependencies}
\usage{
  get_table_reverse(pkg_doc)
}
\arguments{
  \item{pkg_doc}{an object of class
  \code{"HTMLInternalDocument"}}
}
\description{
  Extracts the html table of reverse dependencies
}
\keyword{internal}

