% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{image.brain}
\alias{image.brain}
\title{Print a slice of a 3D image}
\usage{
\method{image}{brain}(x, z = NULL, ...)
}
\description{
Print a slice of a 3D image
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
raw <- read_h5(file)
image(raw)
image(raw, z = 23)
}
