% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_modpac.R
\name{index_repository}
\alias{index_repository}
\title{Index the Contents of a Repository}
\usage{
index_repository(repo = get_repo_location(),
  fields = get_packages_fields(), type = "source", new_pkgs = NULL)
}
\arguments{
\item{repo}{The location of the repository.}

\item{fields}{A vector of fields to use in the \code{"PACKAGES"} files.}

\item{type}{The type of packages in the repository. Only \code{"source"} is
supported at present.}

\item{new_pkgs}{An optional vector of package bundles to recompute metadata
for. This is intended to indicate package bundles that have replaced
existing packages in the index, likely a rare operation for stable repos.}
}
\value{
A data frame of package metadata specified by \code{fields}.
}
\description{
Constructs a data frame of a CRAN-like repository's contents from a
combination of (1) the \code{PACKAGES} file, if it exists; (2) the
\code{DESCRIPTION} files from new package bundles; and (3) any changes due
to removed (or archived) packages.
}
