% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_repo.R
\name{init_repo}
\alias{init_repo}
\title{Initializes a new repo with the given name in the given location}
\usage{
init_repo(repo)
}
\arguments{
\item{repo}{a path}
}
\description{
Initializes a new repo with the given name in the given location
}

