% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_to_repo.R
\name{install_package_to_repo}
\alias{install_package_to_repo}
\title{Installs a Package into a Repository}
\usage{
install_package_to_repo(pkg, repo = get_repo_location(),
  type = "source", repo_name = get_repo_name(), ...)
}
\arguments{
\item{pkg}{A path to a package, e.g. \code{"cranium.tar.gz"}.}

\item{repo}{Repository Directory (by default, this uses the repo location set with "set_repo_location")}

\item{type}{The type of the package. Only \code{"source"} is supported.}

\item{repo_name}{A name to give the repo, will be used by packrat to search for sources in user's options file}

\item{...}{Further arguments passed on to \code{\link{write_modpac}}.}
}
\value{
A Descriptions Table of the packages made available
}
\description{
Installs a Package into a Repository
}
