% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{is_z_curve}
\alias{is_z_curve}
\alias{is_z_curve.tbl_brain}
\title{Identify Curve in yz plane and xz plane}
\usage{
is_z_curve(data, type = "wildtype", threshold.n = 0.9)

\method{is_z_curve}{tbl_brain}(data, type = "wildtype",
  threshold.n = 0.9)
}
\arguments{
\item{data}{a brain image}

\item{type}{wild type or mutant type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded}
}
\description{
Identify Curve in yz plane and xz plane
}
