% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{modify_description}
\alias{modify_description}
\title{Modify the \code{DESCRIPTION} of a Built Package}
\usage{
modify_description(field, value, file)
}
\arguments{
\item{field}{A vector of \code{DESCRIPTION} file fields, e.g.
\code{"Repository"}.}

\item{value}{A vector of string values these fields should take.}

\item{file}{A package bundle, e.g. \code{"cranium_1.0.0.tar.gz"}.}
}
\value{
\code{TRUE} if the bundle was modified correctly and \code{FALSE} otherwise.
}
\description{
This function allows you to modify package metadata inside of an existing
package bundle. Some effort is made to do this as efficiently as possible.
}
