% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{modify_descriptions}
\alias{modify_descriptions}
\title{Modify field in every description}
\usage{
modify_descriptions(field, value)
}
\arguments{
\item{field}{example: Repository}

\item{value}{example: internal}
}
\value{
A new available packages table
}
\description{
Modify field in every description
}
