% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{pipeline_correctionAB}
\alias{pipeline_correctionAB}
\title{Pipeline for error A, error B correction}
\usage{
pipeline_correctionAB(x = ro_brain_yt, B.correct.n = 8)
}
\arguments{
\item{x}{brain data after tidied and PCA reoriented.}

\item{B.correct.n}{number of error B correction being applied. Since error B correction requires several iterations,
suggested number of iteration is 8. Defalt to 8 iterations.}
}
\description{
Detect error A and error B sequentially and make corresponding corrections. Return corected brain data.
}
\examples{


}
