% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs_available.R
\name{pkgs_available}
\alias{pkgs_available}
\title{Checks packages stored at location, wrapper around available.packages from utils}
\usage{
pkgs_available(repo)
}
\arguments{
\item{repo}{a local or external location, relative and absolute paths accepted}
}
\value{
a Descriptions table
}
\description{
Checks packages stored at location, wrapper around available.packages from utils
}
\examples{

# Get Current Repository's Packages
pkgs_available(getOption('repos'))


}
