% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{plot2d}
\alias{plot2d}
\alias{plot2d.tbl_brain}
\title{Plot a 2D projection of a 3D image}
\usage{
plot2d(x, threshold = 0.99, plane = c("x", "z"), ...)

\method{plot2d}{tbl_brain}(x, threshold = 0.99, plane = c("x", "z"), ...)
}
\arguments{
\item{x}{vectors of points to be plotted. Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link[grDevices]{xyz.coords}}
    for details.}

\item{threshold}{value below which points will not be plotted}

\item{plane}{a character vector of length 2 indicating the
plane to project to (x, y, z)}

\item{...}{additional parameters which will be passed to \code{\link{par3d}}, \code{\link{material3d}}
  or \code{decorate3d}.}
}
\description{
Plot a 2D projection of a 3D image
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
tbl_brain <- read_h5(file) \%>\%
  tidy()
plot2d(tbl_brain, plane = c("x", "z"))
plot2d(tbl_brain, plane = c("x", "y"))
plot2d(tbl_brain, plane = c("y", "z"))
}
}
