\name{previous_pkg_html_table}
\alias{previous_pkg_html_table}
\title{Parse HTML Table of an R Package Previous Versions}
\usage{
  previous_pkg_html_table(pkg_name)
}
\arguments{
  \item{pkg_name}{name of an R package in CRAN}
}
\value{
  HTML table of an R package previous versions
}
\description{
  Parse HTML Table of an R Package Previous Versions
}
\examples{
\dontrun{
  # HTML table of previous versions of 'plspm'
  plspm_previous = previous_pkg_html_table("plspm")
}
}
\keyword{internal}

