\name{previous_pkg_url}
\alias{previous_pkg_url}
\title{URL of a package previous versions}
\usage{
  previous_pkg_url(package)
}
\arguments{
  \item{package}{name of an R package in CRAN}
}
\value{
  complete URL of previous package versions
}
\description{
  Returns the URL of a package previous versions in CRAN
}
\examples{
\dontrun{
  # URL of previous versions of 'plspm'
  previous_pkg_url("plspm")
}
}
\keyword{internal}

