% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{qmodel}
\alias{qmodel}
\alias{qmodel.brain}
\alias{qmodel.tbl_brain}
\title{Quadratic model evaluation}
\usage{
qmodel(data, type = "wildtype", threshold.n = 0.9)

\method{qmodel}{brain}(data, type = "wildtype", threshold.n = 0.9)

\method{qmodel}{tbl_brain}(tidy_brain, type = "wildtype",
  threshold.n = 0.9)
}
\arguments{
\item{data}{a brain image or a tidy brain}

\item{type}{wild type or mutant type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded}
}
\description{
Fit quadratic model y=x^2 + x to the data and evaluate how well the model fit the data using adjusted
R-Squared, RMSE, number of signals being captured
}
\examples{
brain <- read_h5(download_brains(tempdir(), pattern = "101"))
qmodel(brain)

#If you have already tidied your brain
tidy_brain <- tidy(brain, type = "wildtype", threshold.n = 0.9)
qmodel(tidy_brain)
}
