% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{read_h5}
\alias{read_h5}
\title{Read HDF5 files}
\usage{
read_h5(file, name = NULL, ...)
}
\arguments{
\item{file}{path to the HDF5 file to be read}

\item{name}{name of the attr}

\item{...}{parameters passed to \code{\link[rhdf5]{h5read}}}
}
\value{
an array of class \code{brain}
}
\description{
Read HDF5 files
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
raw <- read_h5(file)
class(raw)
dim(raw)
}
