% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_pkg}
\alias{remove_pkg}
\title{Remove Packages from a Repository}
\usage{
remove_pkg(pkg, version, repo = get_repo_location(), type = "source")
}
\arguments{
\item{pkg}{A vector of package names to archive, e.g. \code{"cranium"}.}

\item{version}{A vector of valid package versions to archive.}

\item{repo}{The location of the package repository.}

\item{type}{The type of the package. Only \code{"source"} is supported.}
}
\value{
A logical vector indicating which packages were removed
  successfully.
}
\description{
Remove Packages from a Repository
}
