% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{reorient}
\alias{reorient}
\title{PCA}
\usage{
reorient(x, correctionA = FALSE, ...)
}
\arguments{
\item{x}{a \code{\link{tbl_brain}}}

\item{...}{arguments passed to \code{\link[stats]{prcomp}}}
}
\description{
PCA
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  plot3d(tidy_brain)
  tidy_flat <- reorient(tidy_brain)
  plot3d(tidy_flat)

  # compare before and after reorientation
  plot2d(tidy_brain)
  plot2d(tidy_flat)

  # retrieve coefficients
  coef(attr(tidy_flat, "quad_mod"))
}
}
}
