% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_repo.R
\name{repository}
\alias{repository}
\alias{summary.cranium_repository}
\alias{print.cranium_repository}
\alias{print.cranium_repository_summary}
\title{CRAN-Like Repositories}
\usage{
repository(path)

\method{summary}{cranium_repository}(object, ...)

\method{print}{cranium_repository}(x, ...)

\method{print}{cranium_repository_summary}(x, ...)
}
\arguments{
\item{path}{The directory containing the packages.}
}
\value{
An object of class \code{"cranium_repository"}, which points to a
  validated CRAN-like repository.
}
\description{
Identify and work with CRAN-like repositories, including those created by
\code{cranium}.
}
