% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{serve}
\alias{serve}
\title{Serve a CRAN-like Package Repository}
\usage{
serve(repo, port = 8000)
}
\arguments{
\item{repo}{The location of the package repository.}

\item{port}{The port to run the server on.}
}
\description{
Starts a local web server to serve packages from the repository.
}
