% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_packages_fields}
\alias{set_packages_fields}
\title{Modify Default Fields in PACKAGES file}
\usage{
set_packages_fields(fields, append = TRUE)
}
\arguments{
\item{fields}{fields to add (must be present in package descriptions)}

\item{append}{should this be appended to existing fields}
}
\description{
Modify Default Fields in PACKAGES file
}
