% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_repo_location}
\alias{set_repo_location}
\title{set repo location}
\usage{
set_repo_location(x)
}
\arguments{
\item{x}{a relative filepath}
}
\value{
the value of x (invisible)
}
\description{
set repo location
}

