% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_repo_name}
\alias{set_repo_name}
\title{set repo name (optional)}
\usage{
set_repo_name(x)
}
\arguments{
\item{x}{a name for the repo, this will be used by packrat when searching for a repository}
}
\value{
the value of x (invisible)
}
\description{
set repo name (optional)
}
