% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_pkg}
\alias{archive_pkg}
\title{Move a Package to the Repository's Archive}
\usage{
archive_pkg(pkg, version, repo = get_repo_location(), type = "source",
  keep = TRUE)
}
\arguments{
\item{pkg}{The name of the package to archive, e.g. \code{"cranium"}.}

\item{version}{The package version to archive.}

\item{repo}{The location of the package repository.}

\item{type}{The type of the package. Only \code{"source"} is supported.}

\item{keep}{When \code{TRUE}, do not remove the package from the repository
after it has been archived.}
}
\description{
Move a Package to the Repository's Archive
}
