% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_package_to_repo.R
\name{install_package_to_repo}
\alias{install_package_to_repo}
\title{Installs package from repo/github into repository (Not Vectorized)}
\usage{
install_package_to_repo(pkg, repos = getOption("repos"),
  destdir = get_repo_location(), type = "source")
}
\arguments{
\item{pkg}{a package, if using a cranium repository and multiple packages are available, versions can be selected with `_version`}

\item{destdir}{the directory of the repository (NOT the contrib.url)}

\item{repo}{the repo to install from, if this is at a github location, install github will be used}
}
\value{
A Descriptions Table of the packages made available
}
\description{
Installs package from repo/github into repository (Not Vectorized)
}
\examples{
# Install to default apache location
install_package_to_repo('data.table',destdir = '/var/www/html/')
}

