% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{is_errorA}
\alias{is_errorA}
\alias{is_errorA.brain}
\alias{correct_errorA}
\alias{correct_errorA.brain}
\title{PCA alignment error identification}
\usage{
is_errorA(data, type = "wildtype", threshold.n = 0.9)

\method{is_errorA}{brain}(data, type = "wildtype", threshold.n = 0.9)

correct_errorA(data, type = "wildtype", threshold.n = 0.9)

\method{correct_errorA}{brain}(data, type = "wildtype",
  threshold.n = 0.9)
}
\arguments{
\item{data}{a brain image}
}
\description{
Correction A: y-axis flipped
}
\examples{
x <- read_h5(download_brains(tempdir(), pattern = "101"))
is_errorA(x)
}
