% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{is_errorB}
\alias{is_errorB}
\alias{is_errorB.brain}
\title{Identify Error B: Curve in yz plane}
\usage{
is_errorB(data, type = "wildtype", threshold.n = 0.9)

\method{is_errorB}{brain}(data, type = "wildtype", threshold.n = 0.9)
}
\arguments{
\item{data}{a brain image}

\item{type}{wild type or mutant type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded}
}
\description{
Identify Curve in yz plane
}
