% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{is_errorB}
\alias{is_errorB}
\alias{is_errorB.brain}
\alias{is_errorB.tbl_brain}
\alias{correct_errorB.tbl_brain}
\title{Identify Error B: Curve in yz plane}
\usage{
is_errorB(data, type = "wildtype", threshold.n = 0.9,
  ref = "outlier")

\method{is_errorB}{brain}(data, type = "wildtype", threshold.n = 0.9,
  ref = "outlier")

\method{is_errorB}{tbl_brain}(data, type = "wildtype",
  threshold.n = 0.9)

correct_errorB.tbl_brain(data, type = "wildtype", threshold.n = 0.9,
  r_angle_mpt = 1)

correct_errorB.tbl_brain(data, type = "wildtype", threshold.n = 0.9,
  r_angle_mpt = 1)
}
\arguments{
\item{data}{a brain image}

\item{type}{wild type or mutant type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded}
}
\description{
Identify Curve in yz plane
}
