% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{modify_description}
\alias{modify_description}
\title{Modify description of a single zipped package}
\usage{
modify_description(field, value, file)
}
\arguments{
\item{field}{example: Repository}

\item{value}{example: internal}

\item{file}{example: /path/to/cranium_0.3.0.tar.gz}
}
\value{
A new available packages table
}
\description{
Modify description of a single zipped package
}
