% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{plot3d.tbl_brain}
\alias{plot3d.tbl_brain}
\alias{plot3d.brain}
\title{Plot a 3D image of a brain}
\usage{
\method{plot3d}{tbl_brain}(x, ...)

\method{plot3d}{brain}(x, ...)
}
\arguments{
\item{x}{vectors of points to be plotted. Any reasonable way of defining the
    coordinates is acceptable.  See the function \code{\link[grDevices]{xyz.coords}}
    for details.}

\item{...}{additional parameters which will be passed to \code{\link{par3d}}, \code{\link{material3d}}
  or \code{decorate3d}.}
}
\description{
Plot a 3D image of a brain
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  plot3d(tidy_brain)
}
}
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
brain <- read_h5(file)
plot3d(brain)
}
}
