% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{plot_all_sample}
\alias{plot_all_sample}
\title{Plot all samples}
\usage{
plot_all_sample(type = "na", wd = "/Users/czang/Documents/cranium/")
}
\arguments{
\item{type}{wild type or youtoo type of zebrafish embryo brain}
}
\description{
Plot all samples together and output as pdf
}
\examples{
plot_all_sample(type = "wildtype", wd = "/Users/czang/Documents/cranium/")
plot_all_sample(type = "youtoo", wd = "/Users/czang/Documents/cranium/")

}
