% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{process_all_sample}
\alias{process_all_sample}
\title{Read in and tidy all raw samples}
\usage{
process_all_sample(download = FALSE, local_directory = NA,
  type = "youtoo", threshold.n = 0.5, PCA_reorient = TRUE)
}
\arguments{
\item{download}{TRUE or FALSE, If TRUE, you will download data to your current directory. If FALSE, you will need to specify the directory
where you keep you files locally by setting "local_directory".}

\item{local_directory}{If you have samples in your locally, enter the directory. Note than this should be a folder that keeps the samples
of the same sample type such as wild type. You should not keep both wild type and youtoo type in the same folder.}

\item{type}{wild type or youtoo type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded.
Default to 0.5}

\item{PCA_reorient}{TRUE or FALSE. If TRUE, we will tidy and then perform PCA reorientation to your data. If FALSE, we will only tidy your data.}
}
\description{
Read in and tidy all samples. You can choose to perform PCA reorientation or not. Note that this function is only for raw data (untidied data).
Function will output a list containing all samples in the type you decide (wildtype or youtoo), in
the format of dataframes. Processed samples "ro_brain_ls_wt.rds" and "ro_brain_ls_yt.rds" are generated from the
raw data using this function.
}
\examples{
ro_brain_ls_wt <- process_all_sample(download = FALSE, local_directory = NA, type = "wildtype", threshold.n = 0.5, PCA_reorient = TRUE)
ro_brain_ls_yt <- process_all_sample(download = FALSE, local_directory = NA, type = "youtoo", threshold.n = 0.5, PCA_reorient = TRUE)
saveRDS(ro_brain_ls_wt , file = "ro_brain_ls_wt.rds")
saveRDS(ro_brain_ls_yt , file = "ro_brain_ls_yt.rds")
}
