% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{process_all_sample}
\alias{process_all_sample}
\title{Read in and tidy all sample}
\usage{
process_all_sample(type = "youtoo", threshold.n = 0.5,
  PCA_reorient = TRUE)
}
\arguments{
\item{type}{wild type or youtoo type of zebrafish embryo brain}

\item{threshold.n}{threshld for signal frequency. Any signal with frequency below threshold value is exluded.
Default to 0.5}
}
\description{
Read in and tidy all sample. You can choose to process wildtype or youtoo
Function will output a list containing all samples in the type you decide (wildtype or youtoo), in
the format of dataframes. Processed samples "ro_brain_ls_wt.rds" and "ro_brain_ls_yt.rds" are generated from the
raw data using this function.
}
\examples{
ro_brain_ls_wt <- process_all_sample(type = "wildtype", threshold.n = 0.5, PCA_reorient = TRUE)
ro_brain_ls_yt <- process_all_sample(type = "youtoo", threshold.n = 0.5, PCA_reorient = TRUE)
saveRDS(ro_brain_ls_wt , file = "ro_brain_ls_wt.rds")
saveRDS(ro_brain_ls_yt , file = "ro_brain_ls_yt.rds")
}
