% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{repo_descriptions}
\alias{repo_descriptions}
\title{Extract descriptions from a repository}
\usage{
repo_descriptions(repo = get_repo_location())
}
\arguments{
\item{repo}{a path to a repository}
}
\value{
a data.table/data.frame object containing Description info
}
\description{
Extract descriptions from a repository
}
\examples{
repo_descriptions()
}

