% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{tidy.brain}
\alias{tidy.brain}
\title{Tidy 3D brain image data}
\usage{
\method{tidy}{brain}(x, ...)
}
\arguments{
\item{x}{An object to be converted into a tidy data.frame}

\item{...}{extra arguments}
}
\description{
Tidy 3D brain image data
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  class(tidy_brain)
  glimpse(tidy_brain)
}
}
}
