% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radial_bridge.R
\name{tidy.brain}
\alias{tidy.brain}
\title{Tidy 3D brain image data}
\usage{
\method{tidy}{brain}(x, threshold = 0.9, ...)
}
\arguments{
\item{x}{An \code{lm} object created by \code{\link[stats:lm]{stats::lm()}}.}

\item{threshold}{probability below which points will be discarded}

\item{...}{Additional arguments. Not used. Needed to match generic
signature only. \strong{Cautionary note:} Misspelled arguments will be
absorbed in \code{...}, where they will be ignored. If the misspelled
argument has a default value, the default value will be used.
For example, if you pass \code{conf.lvel = 0.9}, all computation will
proceed using \code{conf.level = 0.95}. Additionally, if you pass
\code{newdata = my_tibble} to an \code{\link[=augment]{augment()}} method that does not
accept a \code{newdata} argument, it will use the default value for
the \code{data} argument.}
}
\description{
Tidy 3D brain image data
}
\examples{
file <- "~/Data/barresi/AT_1_Probabilities.h5"
\dontrun{
if (require(dplyr)) {
  tidy_brain <- file \%>\%
    read_h5() \%>\%
    tidy()
  class(tidy_brain)
  glimpse(tidy_brain)
}
}
}
