% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_modpac.R
\name{write_modpac}
\alias{write_modpac}
\title{Writes modified PACKAGES files to allow install.packages to access versions}
\usage{
write_modpac(repo = get_repo_location())
}
\arguments{
\item{repo}{Repository Directory (by default, this uses the repo location set with "set_repo_location")}
}
\value{
A Descriptions Table of the packages made available
}
\description{
Writes modified PACKAGES files to allow install.packages to access versions
}
