% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_modpac.R
\name{write_modpac}
\alias{write_modpac}
\title{Writes modified PACKAGES files to allow install.packages to access versions}
\usage{
write_modpac(repo = get_repo_location())
}
\arguments{
\item{repo}{Repository Directory}

\item{loc}{visible location (External might be http://server/path/to/repo, internal might be file://path/to/repo)}
}
\value{
A Descriptions Table of the packages made available
}
\description{
Writes modified PACKAGES files to allow install.packages to access versions
}

